generalize cross storage
