+++ public/storage.html
@@ -2,7 +2,7 @@
 <html lang="en">
 <head>
     <meta charset="UTF-8">
-    <title>Ethercast SSO Hub</title>
+    <title>Ethercast Storage Hub</title>
     <script src="https://cdn.rawgit.com/zendesk/cross-storage/1.0.0/dist/hub.min.js"
             integrity="sha384-CZHogoIJY4N0uZBmE9EBPymtzS9m3kBfGd+ge65IzoEzMdO8IF1Qx077oWbluuLJ"
             crossorigin="anonymous"></script>
